refactoring: input field 컴포넌트화 [#158]
