+++ frontend/src/shared/components/input/InputField.tsx
@@ -6,14 +6,14 @@ interface InputFieldProps extends React.InputHTMLAttributes<HTMLInputElement> {
 }
 
 const InputField = React.memo(
-  ({ children, value, onChange, ...props }: InputFieldProps) => {
+  ({ children, value, onChange, type = "text", ...props }: InputFieldProps) => {
     return (
-      <label htmlFor={props.id} className="flex items-center gap-2 p-4">
+      <label htmlFor={props.id} className="flex w-full items-center gap-2 p-4">
         <div className="flex-shrink-0">{children}</div>
         <div className="border-border h-3 border-l"></div>
         <input
           id={props.id}
-          type="text"
+          type={type}
           placeholder={props.placeholder}
           className={cn(
             "text-md w-full border-none bg-transparent outline-none",
